requirements.txt
fastapi[standard]
ultralytics
pillow